+++ packages/builders/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @workflow/builders
 
+## 4.0.1-beta.67
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @workflow/core@4.2.0-beta.76
+
 ## 4.0.1-beta.66
 
 ### Patch Changes