+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @workflow/cli
 
+## 4.2.0-beta.76
+
+### Patch Changes
+
+- Updated dependencies [[`ef2218a`](https://github.com/vercel/workflow/commit/ef2218ab22310afa04e4e1709906a86969126e52), [`74c4cdb`](https://github.com/vercel/workflow/commit/74c4cdb6519802e3d56760e971507ffb93bc945b)]:
+  - @workflow/world-local@4.1.0-beta.49
+  - @workflow/world-vercel@4.1.0-beta.47
+  - @workflow/web@4.1.0-beta.47
+  - @workflow/core@4.2.0-beta.76
+  - @workflow/builders@4.0.1-beta.67
+
 ## 4.2.0-beta.75
 
 ### Patch Changes